Incorporated comments: added regression module, updated docstrings, removed mean_percentage_error metric and tests.
