+++ ignite/contrib/metrics/__init__.py
@@ -1,9 +1,3 @@
 
 from ignite.contrib.metrics.average_precision import AveragePrecision
 from ignite.contrib.metrics.roc_auc import ROC_AUC
-from ignite.contrib.metrics.mae import MaximumAbsoluteError
-from ignite.contrib.metrics.fractional_bias import FractionalBias
-from ignite.contrib.metrics.manhattan_distance import ManhattanDistance
-from ignite.contrib.metrics.mean_error import MeanError
-from ignite.contrib.metrics.mean_normalized_bias import MeanNormalizedBias
-from ignite.contrib.metrics.mean_percentage_error import MeanPercentageError
