+++ ignite/contrib/metrics/regression/__init__.py
@@ -1,0 +1,6 @@
+
+from ignite.contrib.metrics.regression.maximum_absolute_error import MaximumAbsoluteError
+from ignite.contrib.metrics.regression.fractional_bias import FractionalBias
+from ignite.contrib.metrics.regression.manhattan_distance import ManhattanDistance
+from ignite.contrib.metrics.regression.mean_error import MeanError
+from ignite.contrib.metrics.regression.mean_normalized_bias import MeanNormalizedBias
