+++ ignite/contrib/metrics/regression/manhattan_distance.py
@@ -13,7 +13,7 @@ class ManhattanDistance(Metric):
     It has been proposed in `Performance Metrics (Error Measures) in Machine Learning Regression, Forecasting and
     Prognostics: Properties and Typology`.
 
-    More details can be found in `here`_.
+    More details can be found in `https://arxiv.org/ftp/arxiv/papers/1809/1809.03006.pdf`.
 
     :math:`\text{MD} = \sum _j^n (A_j - P_j)`
 
@@ -22,8 +22,6 @@ class ManhattanDistance(Metric):
     - `update` must receive output of the form `(y_pred, y)`.
     - `y` and `y_pred` must be of same shape.
 
-    .. _here:
-        https://arxiv.org/ftp/arxiv/papers/1809/1809.03006.pdf
     """
     def reset(self):
         self._sum_of_errors = 0.0
