+++ ignite/contrib/metrics/regression/maximum_absolute_error.py
@@ -10,17 +10,15 @@ class MaximumAbsoluteError(Metric):
     It has been proposed in `Performance Metrics (Error Measures) in Machine Learning Regression, Forecasting and
     Prognostics: Properties and Typology`.
 
-    More details can be found in `here`_.
+    More details can be found in `https://arxiv.org/ftp/arxiv/papers/1809/1809.03006.pdf`.
 
-    :math:`\text{MaxAE} = \max_{j=1,n} \left(|A_j-P_j\right|)`
+    :math:`\text{MaxAE} = \max_{j=1,n} \left( \lvert A_j-P_j \rvert \right)`
 
     Where, :math:`A_j` is the ground truth and :math:`P_j` is the predicted value.
 
     - `update` must receive output of the form `(y_pred, y)`.
     - `y` and `y_pred` must be of same shape.
 
-    .. _here:
-        https://arxiv.org/ftp/arxiv/papers/1809/1809.03006.
     """
 
     def reset(self):
